Migrate Component types to dc_bundle
Fixes Migrate Component types to dc_bundle #1454
