+++ crates/dc_bundle/src/legacy_definition/view/component.rs
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2024 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use crate::legacy_definition::view::text_style::StyledTextRun;
+use crate::legacy_definition::view::view_style::ViewStyle;
+use serde::{Deserialize, Serialize};
+
+/// Figma component properties can be "overridden" in the UI. These overrides
+/// are applied in the node tree we get back from Figma, so we don't have to
+/// worry about them, unless we present a variant of a component as a result of
+/// an interaction. In that case, we need to figure out what has been overridden
+/// and apply it to the variant. This struct keeps track of those overrides that
+/// need to be applied to variant instances.
+#[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
+pub struct ComponentOverrides {
+    pub style: Option<ViewStyle>,
+    pub data: ComponentContentOverride,
+}
+
+#[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
+pub enum ComponentContentOverride {
+    None,
+    Text { content: String, res_name: Option<String> },
+    StyledText { content: Vec<StyledTextRun>, res_name: Option<String> },
+}
+
+/// Details on the Figma component that this view is an instance of.
+#[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
+pub struct ComponentInfo {
+    pub id: String,
+    pub name: String,
+    pub component_set_name: String,
+    // Currently we only figure out overrides for the root of a component, but soon
+    // we will want to compute them for the whole tree.
+    pub overrides: Option<ComponentOverrides>,
+}
