+++ crates/figma_import/src/reflection.rs
@@ -163,10 +163,12 @@ pub fn registry() -> serde_reflection::Result<serde_reflection::Registry> {
         .trace_type::<dc_bundle::legacy_definition::element::path::Path>(&samples)
         .expect("couldn't trace Path");
     tracer
-        .trace_type::<crate::toolkit_schema::ComponentContentOverride>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::view::component::ComponentContentOverride>(
+            &samples,
+        )
         .expect("couldn't trace ComponentContentOverride");
     tracer
-        .trace_type::<crate::toolkit_schema::ComponentInfo>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::view::component::ComponentInfo>(&samples)
         .expect("couldn't trace ComponentInfo");
     tracer
         .trace_type::<dc_bundle::legacy_definition::layout::positioning::OverflowDirection>(
